fix(scrobble): host delay gate in qbz-app for unit testing
Extract scrobble_delay_secs so cargo test -p qbz-app covers unknown
duration without compiling the Slint binary.
